+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 # Unreleased
 
 * Fix cookie banner issue (IE10) ([PR #2231](https://github.com/alphagov/govuk_publishing_components/pull/2231))
+* Extend layout for public with account components ([PR #2255](https://github.com/alphagov/govuk_publishing_components/pull/2255))
 
 # 25.2.3